FACE CLASSIFICATION
SVMs are a powerful class of supervised learning algorithms for classification and regression problems. In the context of classification, SVMs can be viewed as maximum margin linear classifiers.
The SVM uses an objective which explicitly encourages low out-of-sample error (good generalization performance). The D dimensional data are divided into classes by maximizing the margin between the hyperplanes for the classes.